requirements.txt
torch
torchvision
streamlit
numpy
opencv-python
pillow
matplotlib
seaborn
scikit-learn